Split check result count in a new component
